Created Coming Soon page
Created the coming soon page with an data from an API that provides the movies that are coming out soon. JSON data is provided as a way to preserve the number of API calls. Did find out that the only way that the only way that the page loads with the local JSON data is if you run or simulate a local server otherwise you run into a CORS error that doesn't allow the js script to read a file. I was doing this unknowingly by using a VSCode extension that allows me to preview html pages called live preview.