Coming Soon and showtimes are complete
Added the showtimes page. Once again the only way to view this without the API being active is to simulate a server. There is more info on that in the readme. Also made small changes to both table generating JavaScript files. Currently each theater has it's own table and all poster images have a meta tag.